+++ Form/Action/DuplicateAction.php
@@ -1,9 +1,9 @@
 <?php declare(strict_types=1);
 
-namespace Yireo\LokiAdminComponents\Form\Action;
+namespace Loki\AdminComponents\Form\Action;
 
 use Magento\Framework\Message\Manager;
-use Yireo\LokiAdminComponents\Component\Form\FormRepository;
+use Loki\AdminComponents\Component\Form\FormRepository;
 
 class DuplicateAction implements ActionInterface
 {