+++ Form/Field/FieldTypeInterface.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Yireo\LokiAdminComponents\Form\Field;
+namespace Loki\AdminComponents\Form\Field;
 
 interface FieldTypeInterface
 {